chore: Rename to lightwave #1796
