+++ .github/workflows/release-wave.yml
@@ -157,14 +157,14 @@ jobs:
           packages_dir: university/dist
           password: ${{ secrets.PYPI_UNIVERSITY_TOKEN }}
 
-      - name: Publish to PyPI - wavelite
+      - name: Publish to PyPI - lightwave
         uses: pypa/gh-action-pypi-publish@master
         with:
-          packages_dir: py/h2o_wavelite/dist
+          packages_dir: py/h2o_lightwave/dist
           password: ${{ secrets.PYPI_WAVELITE_TOKEN }}
 
-      - name: Publish to PyPI - wavelite_web
+      - name: Publish to PyPI - lightwave_web
         uses: pypa/gh-action-pypi-publish@master
         with:
-          packages_dir: py/h2o_wavelite_web/dist
+          packages_dir: py/h2o_lightwave_web/dist
           password: ${{ secrets.PYPI_WAVELITE_WEB_TOKEN }}
