+++ Makefile
@@ -205,11 +205,11 @@ publish-vsc-extension: ## Publish VS Code extension
 publish-university:
 	cd university && $(MAKE) publish
 	
-publish-wavelite:
+publish-lightwave:
 	cd ui && npm ci && npm run build
 	cd py && $(MAKE) setup
-	cd py && $(MAKE) build-wavelite
-	cd py && $(MAKE) build-wavelite-web
+	cd py && $(MAKE) build-lightwave
+	cd py && $(MAKE) build-lightwave-web
 	
 .PHONY: tag
 tag: ## Bump version and tag
