+++ py/.gitignore
@@ -133,4 +133,3 @@ h2o_wave/h2o_wave/metadata.py
 # Wave ML
 /h2o_wave_ml
 examples/_tour_apps_tmp
-h2o_wavelite/examples/index.html
