+++ py/h2o_lightwave/examples/hello_fastapi.py
@@ -1,14 +1,10 @@
-import os
-
 from fastapi import FastAPI, WebSocket, WebSocketDisconnect
-from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
-from h2o_wavelite import Q, ui, wave_serve
-from h2o_wavelite_web import get_web_files, web_directory
-from jinja2 import Environment, FileSystemLoader
+from h2o_lightwave import Q, ui, wave_serve
+from h2o_lightwave_web import web_directory
 
 
-# Wavelite callback function.
+# Lightwave callback function.
 async def serve(q: Q):
     # Paint our UI on the first page visit.
     if not q.client.initialized:
@@ -32,13 +28,13 @@ async def serve(q: Q):
     await q.page.save()
 
 
-# Run: uvicorn hello_fastapi_custom_index:app.
+# Run: uvicorn hello_fastapi:app.
 # FastAPI boilerplate.
 app = FastAPI()
 
 
 # FastAPI: WebSocket must be registered before index.html handler.
-@app.websocket("/custom_socket/")
+@app.websocket("/_s/")
 async def ws(ws: WebSocket):
     try:
         await ws.accept()
@@ -47,18 +43,4 @@ async def ws(ws: WebSocket):
     except WebSocketDisconnect:
         print('Client disconnected')
 
-
-# Where do we want to serve our assets from?
-assets_path = '/custom/assets/path'
-curr_dir = os.path.dirname(os.path.abspath(__file__))
-
-# Prepare our custom index.html and inject required JS files.
-# Jinja is used for convenience, you can use any templating engine.
-template = Environment(loader=FileSystemLoader(curr_dir)).get_template("index_template.html")
-
-@app.get("/")
-async def get():
-    # Inject JS files into the template. Accepts a path prefix if needed.
-    return HTMLResponse(template.render(wave_files=get_web_files(assets_path)))
-
-app.mount(assets_path, StaticFiles(directory=web_directory, html=True), name=assets_path)
+app.mount("/", StaticFiles(directory=web_directory, html=True), name="/")
