+++ py/h2o_lightwave_web/MANIFEST.in
@@ -1,0 +1,1 @@
+recursive-include h2o_lightwave_web/www *
