+++ py/h2o_lightwave_web/Makefile
@@ -1,7 +1,7 @@
 .PHONY: build
 build: purge
-	cp -R ../../ui/build h2o_wavelite_web/www
+	cp -R ../../ui/build h2o_lightwave_web/www
 	../venv/bin/python setup.py bdist_wheel
 
 purge:
-	rm -rf h2o_wavelite_web/www
+	rm -rf h2o_lightwave_web/www
