+++ py/h2o_lightwave_web/README.md
@@ -1,0 +1,1 @@
+Web assets for H2O Lightwave.
