--- py/h2o_wavelite_web/.gitignore
@@ -1,1 +1,0 @@
-h2o_wavelite_web/www
