--- py/h2o_wavelite_web/MANIFEST.in
@@ -1,1 +1,0 @@
-recursive-include h2o_wavelite_web/www *
