--- py/h2o_wavelite_web/README.md
@@ -1,1 +1,0 @@
-Web assets for H2O Wavelite.
